fix(convert.py): allow multiple user to convert simultaneously
We shared the socket between every py3dtiles instance. First, I don't
know how zmq behaves in this case, but it's probably bad. Secondly, it
prevents 2 different users to use py3dtiles, because the second user
won't have permission on the socket.

Also, the socket was never removed, so the second user couldn't convert
at all until the socket was cleaned (a reboot does that when the socket
is in a tmpfs).

Fixes #169