ci(docker): use dind docker image to build docker image
I could insert a meme here (you know which one). What
this means is that we previously build our docker image in a virtualbox
executor (so a vm) on Oslandia infrastructure. Now that we have the
gitlab.com infra, things are not done that way and we must use
docker-in-docker images.